refactor(dashboard-scene): extract getQueryRunnerFor into its own module
`utils/utils.ts` imports DashboardScene, so anything importing getQueryRunnerFor
from it pulls the whole scene graph in behind the function. That is a cycle
waiting to happen for any module DashboardScene itself reaches.

The function depends on nothing but @grafana/scenes, so move it to its own
module and point the call sites at it. No behaviour change: the body is
unchanged and every change here is an import path.
