+++ public/app/features/dashboard-scene/inspect/HelpWizard/utils.ts
@@ -16,7 +16,8 @@ import { GrafanaQueryType } from 'app/plugins/datasource/grafana/types';
 
 import { type DashboardGridItem } from '../../scene/layout-default/DashboardGridItem';
 import { gridItemToPanel, vizPanelToPanel } from '../../serialization/transformSceneToSaveModel';
-import { getQueryRunnerFor, isLibraryPanel } from '../../utils/utils';
+import { getQueryRunnerFor } from '../../utils/getQueryRunnerFor';
+import { isLibraryPanel } from '../../utils/utils';
 
 import { type Randomize, randomizeData } from './randomizer';
