+++ public/app/features/dashboard-scene/mutation-api/commands/updatePanel.ts
@@ -17,7 +17,7 @@ import { AutoGridItem } from '../../scene/layout-auto-grid/AutoGridItem';
 import { PanelTimeRange } from '../../scene/panel-timerange/PanelTimeRange';
 import { getUpdatedHoverHeader } from '../../scene/panel-timerange/utils';
 import { getElements, panelQueryKindToSceneQuery } from '../../serialization/layoutSerializers/utils';
-import { getQueryRunnerFor } from '../../utils/utils';
+import { getQueryRunnerFor } from '../../utils/getQueryRunnerFor';
 import { getVizPanelKeyForPanelId } from '../../utils/utils-panels';
 
 import { serializeResultLayoutItem } from './panelSerialization';
