+++ public/app/features/dashboard-scene/scene/layout-default/DashboardGridItem.test.tsx
@@ -23,8 +23,8 @@ setPluginImportUtils({
 });
 
 const mockGetQueryRunnerFor = jest.fn();
-jest.mock('../../utils/utils', () => ({
-  ...jest.requireActual('../../utils/utils'),
+jest.mock('../../utils/getQueryRunnerFor', () => ({
+  ...jest.requireActual('../../utils/getQueryRunnerFor'),
   getQueryRunnerFor: jest.fn().mockImplementation(() => mockGetQueryRunnerFor()),
 }));
