+++ public/app/features/dashboard-scene/scene/panel-timerange/PanelTimeRangeDrawer.tsx
@@ -23,7 +23,8 @@ import {
 } from '@grafana/ui';
 
 import { getQuickOptions } from '../../../../../../packages/grafana-ui/src/components/DateTimePickers/options';
-import { getDashboardSceneFor, getQueryRunnerFor } from '../../utils/utils';
+import { getQueryRunnerFor } from '../../utils/getQueryRunnerFor';
+import { getDashboardSceneFor } from '../../utils/utils';
 
 import { PanelTimeRange } from './PanelTimeRange';
