+++ public/app/features/dashboard-scene/sharing/public-dashboards/utils.ts
@@ -4,7 +4,7 @@ import { type VizPanel } from '@grafana/scenes';
 import { supportedDatasources } from 'app/features/dashboard/components/ShareModal/SharePublicDashboard/SupportedPubdashDatasources';
 
 import { type DashboardScene } from '../../scene/DashboardScene';
-import { getQueryRunnerFor } from '../../utils/utils';
+import { getQueryRunnerFor } from '../../utils/getQueryRunnerFor';
 
 export const getUnsupportedDashboardDatasources = async (types: string[]): Promise<string[]> => {
   let unsupportedDS = new Set<string>();
