+++ public/app/features/dashboard-scene/utils/PanelModelCompatibilityWrapper.ts
@@ -2,7 +2,7 @@ import { type PanelModel } from '@grafana/data';
 import { SceneDataTransformer, type VizPanel } from '@grafana/scenes';
 import { type DataSourceRef, type DataTransformerConfig } from '@grafana/schema';
 
-import { getQueryRunnerFor } from './utils';
+import { getQueryRunnerFor } from './getQueryRunnerFor';
 import { getPanelIdForVizPanel } from './utils-panels';
 
 export class PanelModelCompatibilityWrapper implements PanelModel {
