+++ public/app/features/dashboard-scene/utils/utils.ts
@@ -238,24 +238,6 @@ export function getCurrentValueForOldIntervalModel(variable: IntervalVariableMod
   return intervals[0];
 }
 
-export function getQueryRunnerFor(sceneObject: SceneObject | undefined): SceneQueryRunner | undefined {
-  if (!sceneObject) {
-    return undefined;
-  }
-
-  const dataProvider = sceneObject.state.$data ?? sceneObject.parent?.state.$data;
-
-  if (dataProvider instanceof SceneQueryRunner) {
-    return dataProvider;
-  }
-
-  if (dataProvider instanceof SceneDataTransformer) {
-    return getQueryRunnerFor(dataProvider);
-  }
-
-  return undefined;
-}
-
 export function getDashboardSceneFor(sceneObject: SceneObject): DashboardScene {
   const root = sceneObject.getRoot();
